refactor(MovieItemsSkeleton): 스켈레톤 컴포넌트 분리
