+++ src/components/MovieItems/MovieItems.ts
@@ -1,21 +1,9 @@
-import OPTIONS from '../../constants/OPTIONS';
 import ResponseData from '../../interfaces/ResponseData';
 import MovieData from '../../interfaces/MovieData';
 import { starFilled } from '../../resources';
 import { MOVIE_POSTER_URL } from '../../constants/DTO';
 
 const MovieItems = {
-  createSkeleton(): HTMLUListElement {
-    const movieItems = document.createElement('ul');
-    movieItems.classList.add('item-list');
-
-    Array.from({ length: OPTIONS.movieItemCount }).forEach(() => {
-      movieItems.appendChild(this.createMovieItemSkeleton());
-    });
-
-    return movieItems;
-  },
-
   replaceSkeletons(movieItems: HTMLUListElement, respondData: ResponseData) {
     const itemCards = movieItems.querySelectorAll('li');
     itemCards.forEach((item, index) => this.replaceSkeleton(item, respondData.results[index]));
@@ -72,62 +60,6 @@ const MovieItems = {
     star.alt = '별점';
 
     return star;
-  },
-
-  createMovieItemSkeleton() {
-    const movieItem = document.createElement('li');
-
-    const movieItemLink = this.createMovieItemLink(this.createMovieItemCardSkeleton());
-
-    movieItem.appendChild(movieItemLink);
-
-    return movieItem;
-  },
-
-  createMovieItemLink(movieItemCard: HTMLElement) {
-    const movieItemLink = document.createElement('a');
-
-    movieItemLink.setAttribute('href', '#');
-
-    movieItemLink.appendChild(movieItemCard);
-
-    return movieItemLink;
-  },
-
-  createMovieItemCardSkeleton() {
-    const movieItemCardSkeleton = document.createElement('div');
-
-    movieItemCardSkeleton.classList.add('item-card');
-
-    movieItemCardSkeleton.appendChild(this.createMovieItemThumbnailSkeleton());
-    movieItemCardSkeleton.appendChild(this.createMovieItemTitleSkeleton());
-    movieItemCardSkeleton.appendChild(this.createMovieItemScoreSkeleton());
-
-    return movieItemCardSkeleton;
-  },
-
-  createMovieItemThumbnailSkeleton() {
-    const movieItemThumbnail = document.createElement('div');
-
-    movieItemThumbnail.classList.add('item-thumbnail', 'skeleton');
-
-    return movieItemThumbnail;
-  },
-
-  createMovieItemTitleSkeleton() {
-    const movieItemTitle = document.createElement('div');
-
-    movieItemTitle.classList.add('item-title', 'skeleton');
-
-    return movieItemTitle;
-  },
-
-  createMovieItemScoreSkeleton() {
-    const movieItemScore = document.createElement('div');
-
-    movieItemScore.classList.add('item-score', 'skeleton');
-
-    return movieItemScore;
   },
 };
